requirements.txt
# Production Telegram Sync Bot Requirements
python-telegram-bot>=20.0
telethon>=1.32.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.8
aiosqlite>=0.19.0
apscheduler>=3.10.0
requests>=2.31.0
openai>=1.0.0 gspread==5.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
